ActiveRecord >= 6.1 compatibility (#61)
* Update rubocop to be a bit more lenient with class/method length

* Update gemspec to ensure ActiveRecord < 6.2

* Add compatibility for ActiveRecord >= 6.1

* Re-remove Gemfile.lock

* Add tests for add_index and remove_index methods

* Fix specs

* Fix add_index for ruby3/AR6

* Fix test matrix since activerecord 6.0 is incompatible with ruby 3

* v0.3.0

* Add custom matcher replacing matching method

* Update script to automatically test against all gemfiles in the gemfiles directory

+++ lib/active_record/connection_adapters/mysql2_ghost_adapter.rb
@@ -55,14 +55,47 @@ module ActiveRecord
         end
       end
 
-      def add_index(table_name, column_name, options = {})
-        index_name, index_type, index_columns, index_options = add_index_options(table_name, column_name, options)
-        execute "ALTER TABLE #{quote_table_name(table_name)} ADD #{index_type} INDEX #{quote_column_name(index_name)} (#{index_columns})#{index_options}" # rubocop:disable Layout/LineLength
-      end
+      if Gem.loaded_specs['activerecord'].version >= Gem::Version.new('6.1')
+        def add_index(table_name, column_name, **options)
+          index, algorithm, if_not_exists = add_index_options(table_name, column_name, **options)
+          return if if_not_exists && index_exists?(table_name, column_name, name: index.name)
 
-      def remove_index(table_name, options = {})
-        index_name = index_name_for_remove(table_name, options)
-        execute "ALTER TABLE #{quote_table_name(table_name)} DROP INDEX #{quote_column_name(index_name)}"
+          index_type = index.type&.to_s&.upcase || (index.unique ? 'UNIQUE' : nil)
+
+          sql = build_add_index_sql(
+            table_name, quoted_columns(index), index.name,
+            index_type: index_type,
+            using: index.using,
+            algorithm: algorithm
+          )
+
+          execute sql
+        end
+
+        def remove_index(table_name, column_name = nil, **options)
+          return if options[:if_exists] && !index_exists?(table_name, column_name, **options)
+
+          index_name = index_name_for_remove(table_name, column_name, options)
+          execute "ALTER TABLE #{quote_table_name(table_name)} DROP INDEX #{quote_column_name(index_name)}"
+        end
+      else
+        def add_index(table_name, column_name, options = {})
+          index_name, index_type, index_columns, _index_options = add_index_options(table_name, column_name, options)
+
+          sql = build_add_index_sql(
+            table_name, index_columns, index_name,
+            index_type: index_type&.upcase,
+            using: options[:using]
+          )
+
+          execute sql
+        end
+
+        def remove_index(table_name, options = {})
+          options = { column: options } unless options.is_a?(Hash)
+          index_name = index_name_for_remove(table_name, options)
+          execute "ALTER TABLE #{quote_table_name(table_name)} DROP INDEX #{quote_column_name(index_name)}"
+        end
       end
 
       private
@@ -110,6 +143,25 @@ module ActiveRecord
       def schema_migration_update?(sql)
         INSERT_SCHEMA_MIGRATION_PATTERN =~ sql ||
           DROP_SCHEMA_MIGRATION_PATTERN =~ sql
+      end
+
+      def build_add_index_sql(table_name, column_names, index_name, # rubocop:disable Metrics/ParameterLists
+                              index_type: nil, using: nil, algorithm: nil)
+        sql = %w[ALTER TABLE]
+        sql << quote_table_name(table_name)
+        sql << 'ADD'
+        sql << index_type
+        sql << 'INDEX'
+        sql << quote_column_name(index_name)
+        sql << "USING #{using}" if using
+        sql << "(#{column_names})"
+        sql << algorithm
+
+        sql.compact.join(' ').gsub(/\s+/, ' ')
+      end
+
+      def quoted_columns(index)
+        index.columns.is_a?(String) ? index.columns : quoted_columns_for_index(index.columns, index.column_options)
       end
     end
   end
